Checking for right number of parameters

+++ StringTools.hpp
@@ -6,7 +6,7 @@
 /*   By: fde-capu <fde-capu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/07 01:43:08 by fde-capu          #+#    #+#             */
-/*   Updated: 2022/03/28 14:46:12 by fde-capu         ###   ########.fr       */
+/*   Updated: 2022/03/28 16:33:08 by fde-capu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,7 @@ class StringTools
 		bool valid_file_name(std::string) const;
 		bool isAllInSet(std::string, std::string) const;
 		std::string nth_word(std::string, int) const;
-		int word_count(std::string) const;
+		size_t word_count(std::string) const;
 };
 
 //std::ostream & operator<< (std::ostream & o, StringTools const & i);
@@ -110,6 +110,6 @@ std::string itoa(int);
 bool valid_file_name(std::string);
 bool isAllInSet(std::string, std::string);
 std::string nth_word(std::string, int);
-int word_count(std::string);
+size_t word_count(std::string);
 
 #endif
